improved CLI
- made preview optional
- store_true, store_false for boolean inputs
- removed dublicated 'dest' parameters

+++ stitching/cli/stitch.py
@@ -31,7 +31,6 @@ parser.add_argument(
     help="Resolution for image registration step. "
     "The default is %s Mpx" % ImageHandler.DEFAULT_MEDIUM_MEGAPIX,
     type=float,
-    dest="medium_megapix",
 )
 parser.add_argument(
     "--detector",
@@ -41,15 +40,13 @@ parser.add_argument(
     "The default is '%s'." % FeatureDetector.DEFAULT_DETECTOR,
     choices=FeatureDetector.DETECTOR_CHOICES.keys(),
     type=str,
-    dest="detector",
 )
 parser.add_argument(
     "--nfeatures",
     action="store",
     default=500,
-    help="Type of features used for images matching. " "The default is 500.",
+    help="Type of features used for images matching. The default is 500.",
     type=int,
-    dest="nfeatures",
 )
 parser.add_argument(
     "--matcher_type",
@@ -59,7 +56,6 @@ parser.add_argument(
     "The default is '%s'." % FeatureMatcher.DEFAULT_MATCHER,
     choices=FeatureMatcher.MATCHER_CHOICES,
     type=str,
-    dest="matcher_type",
 )
 parser.add_argument(
     "--range_width",
@@ -67,16 +63,11 @@ parser.add_argument(
     default=FeatureMatcher.DEFAULT_RANGE_WIDTH,
     help="uses range_width to limit number of images to match with.",
     type=int,
-    dest="range_width",
 )
 parser.add_argument(
     "--try_use_gpu",
-    action="store",
-    default=False,
-    help="Try to use CUDA. The default value is no. "
-    "All default values are for CPU mode.",
-    type=bool,
-    dest="try_use_gpu",
+    action="store_true",
+    help="Try to use CUDA",
 )
 parser.add_argument(
     "--match_conf",
@@ -84,7 +75,6 @@ parser.add_argument(
     help="Confidence for feature matching step. "
     "The default is 0.3 for ORB and 0.65 for other feature types.",
     type=float,
-    dest="match_conf",
 )
 parser.add_argument(
     "--confidence_threshold",
@@ -93,7 +83,6 @@ parser.add_argument(
     help="Threshold for two images are from the same panorama confidence. "
     "The default is '%s'." % Subsetter.DEFAULT_CONFIDENCE_THRESHOLD,
     type=float,
-    dest="confidence_threshold",
 )
 parser.add_argument(
     "--matches_graph_dot_file",
@@ -101,7 +90,6 @@ parser.add_argument(
     default=Subsetter.DEFAULT_MATCHES_GRAPH_DOT_FILE,
     help="Save matches graph represented in DOT language to <file_name> file.",
     type=str,
-    dest="matches_graph_dot_file",
 )
 parser.add_argument(
     "--estimator",
@@ -111,7 +99,6 @@ parser.add_argument(
     "The default is '%s'." % CameraEstimator.DEFAULT_CAMERA_ESTIMATOR,
     choices=CameraEstimator.CAMERA_ESTIMATOR_CHOICES.keys(),
     type=str,
-    dest="estimator",
 )
 parser.add_argument(
     "--adjuster",
@@ -121,7 +108,6 @@ parser.add_argument(
     "The default is '%s'." % CameraAdjuster.DEFAULT_CAMERA_ADJUSTER,
     choices=CameraAdjuster.CAMERA_ADJUSTER_CHOICES.keys(),
     type=str,
-    dest="adjuster",
 )
 parser.add_argument(
     "--refinement_mask",
@@ -135,7 +121,6 @@ parser.add_argument(
     "parameter then the respective flag is ignored."
     "" % CameraAdjuster.DEFAULT_REFINEMENT_MASK,
     type=str,
-    dest="refinement_mask",
 )
 parser.add_argument(
     "--wave_correct_kind",
@@ -145,7 +130,6 @@ parser.add_argument(
     "The default is '%s'" % WaveCorrector.DEFAULT_WAVE_CORRECTION,
     choices=WaveCorrector.WAVE_CORRECT_CHOICES.keys(),
     type=str,
-    dest="wave_correct_kind",
 )
 parser.add_argument(
     "--warper_type",
@@ -154,7 +138,6 @@ parser.add_argument(
     help="Warp surface type. The default is '%s'." % Warper.DEFAULT_WARP_TYPE,
     choices=Warper.WARP_TYPE_CHOICES,
     type=str,
-    dest="warper_type",
 )
 parser.add_argument(
     "--low_megapix",
@@ -163,18 +146,23 @@ parser.add_argument(
     help="Resolution for seam estimation and exposure estimation step. "
     "The default is %s Mpx." % ImageHandler.DEFAULT_LOW_MEGAPIX,
     type=float,
-    dest="low_megapix",
 )
 parser.add_argument(
     "--crop",
-    action="store",
-    default=Cropper.DEFAULT_CROP,
+    action="store_true",
     help="Crop black borders around images caused by warping using the "
     "largest interior rectangle. "
     "Default is '%s'." % Cropper.DEFAULT_CROP,
-    type=bool,
+)
+parser.add_argument(
+    "--no-crop",
+    action="store_false",
+    help="Don't Crop black borders around images caused by warping using the "
+    "largest interior rectangle. "
+    "Default is '%s'." % (not Cropper.DEFAULT_CROP),
     dest="crop",
 )
+parser.set_defaults(crop=Cropper.DEFAULT_CROP)
 parser.add_argument(
     "--compensator",
     action="store",
@@ -183,7 +171,6 @@ parser.add_argument(
     "The default is '%s'." % ExposureErrorCompensator.DEFAULT_COMPENSATOR,
     choices=ExposureErrorCompensator.COMPENSATOR_CHOICES.keys(),
     type=str,
-    dest="compensator",
 )
 parser.add_argument(
     "--nr_feeds",
@@ -191,7 +178,6 @@ parser.add_argument(
     default=ExposureErrorCompensator.DEFAULT_NR_FEEDS,
     help="Number of exposure compensation feed.",
     type=np.int32,
-    dest="nr_feeds",
 )
 parser.add_argument(
     "--block_size",
@@ -200,7 +186,6 @@ parser.add_argument(
     help="BLock size in pixels used by the exposure compensator. "
     "The default is '%s'." % ExposureErrorCompensator.DEFAULT_BLOCK_SIZE,
     type=np.int32,
-    dest="block_size",
 )
 parser.add_argument(
     "--finder",
@@ -210,7 +195,6 @@ parser.add_argument(
     "The default is '%s'." % SeamFinder.DEFAULT_SEAM_FINDER,
     choices=SeamFinder.SEAM_FINDER_CHOICES.keys(),
     type=str,
-    dest="finder",
 )
 parser.add_argument(
     "--final_megapix",
@@ -219,7 +203,6 @@ parser.add_argument(
     help="Resolution for compositing step. Use -1 for original resolution. "
     "The default is %s" % ImageHandler.DEFAULT_FINAL_MEGAPIX,
     type=float,
-    dest="final_megapix",
 )
 parser.add_argument(
     "--blender_type",
@@ -228,7 +211,6 @@ parser.add_argument(
     help="Blending method. The default is '%s'." % Blender.DEFAULT_BLENDER,
     choices=Blender.BLENDER_CHOICES,
     type=str,
-    dest="blender_type",
 )
 parser.add_argument(
     "--blend_strength",
@@ -237,7 +219,6 @@ parser.add_argument(
     help="Blending strength from [0,100] range. "
     "The default is '%s'." % Blender.DEFAULT_BLEND_STRENGTH,
     type=np.int32,
-    dest="blend_strength",
 )
 parser.add_argument(
     "--timelapse",
@@ -248,7 +229,11 @@ parser.add_argument(
     "The default is '%s'." % Timelapser.DEFAULT_TIMELAPSE,
     choices=Timelapser.TIMELAPSE_CHOICES,
     type=str,
-    dest="timelapse",
+)
+parser.add_argument(
+    "--preview",
+    action="store_true",
+    help="Opens a preview window with the stitched result",
 )
 parser.add_argument(
     "--output",
@@ -256,7 +241,6 @@ parser.add_argument(
     default="result.jpg",
     help="The default is 'result.jpg'",
     type=str,
-    dest="output",
 )
 
 __doc__ += "\n" + parser.format_help()
@@ -269,6 +253,7 @@ def main():
     # Extract In- and Output
     img_names = args_dict.pop("img_names")
     img_names = [cv.samples.findFile(img_name) for img_name in img_names]
+    preview = args_dict.pop("preview")
     output = args_dict.pop("output")
 
     stitcher = Stitcher(**args_dict)
@@ -276,12 +261,13 @@ def main():
 
     cv.imwrite(output, panorama)
 
-    zoom_x = 600.0 / panorama.shape[1]
-    preview = cv.resize(panorama, dsize=None, fx=zoom_x, fy=zoom_x)
+    if preview:
+        zoom_x = 600.0 / panorama.shape[1]
+        preview = cv.resize(panorama, dsize=None, fx=zoom_x, fy=zoom_x)
 
-    cv.imshow(output, preview)
-    cv.waitKey()
-    cv.destroyAllWindows()
+        cv.imshow(output, preview)
+        cv.waitKey()
+        cv.destroyAllWindows()
 
 
 if __name__ == "__main__":
